src/backend/requirements.txt
asgiref==3.8.1
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.8
Django==5.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dj-rest-auth==7.0.1
django-allauth==65.12.1
cryptography>=42,<45
et-xmlfile==1.1.0
filelock==3.16.1
fonttools==4.47.2
formulas==1.2.8
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
mlxtend==0.23.1
numpy==1.26.3
numpy-financial==1.0.0
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
platformdirs==4.3.6
pyarrow==15.0.0
PyJWT==2.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
schedula==1.5.49
scikit-learn==1.4.0
scipy==1.12.0
secure-smtplib==0.1.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.2.0
tzdata==2023.4
virtualenv==20.26.6
psycopg2-binary==2.9.9
watchfiles~=0.21
django-cors-headers==4.9.0
Django==5.1.4
gunicorn==23.0.0
dj-database-url==2.3.0
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.0
watchdog==6.0.0
coverage==7.6.10
unittest-xml-reporting==3.2.0
ruff==0.14.14